Renamed messaging module
