--- src/iotdevicesimulator/__assets__/configs.cfg
@@ -1,12 +1,0 @@
-oracle: {
-    dsn: "wldbase:1521/wla"
-    user: "lewcha"
-    password: "Lucid+Barrel8Valley"
-}
-iot_core: {
-    endpoint: "a10mem0twl4qxt-ats.iot.eu-west-2.amazonaws.com"
-    port: 8883
-    cert_path: "C:/Users/lewcha/OneDrive - UKCEH/Documents/FDRI/projects/iot-device-simulator/src/iotdevicesimulator/__assets__/.certs/fdri_swarm.cert.pem"
-    key_path: "C:/Users/lewcha/OneDrive - UKCEH/Documents/FDRI/projects/iot-device-simulator/src/iotdevicesimulator/__assets__/.certs/fdri_swarm.private.key"
-    ca_cert_path: "C:/Users/lewcha/OneDrive - UKCEH/Documents/FDRI/projects/iot-device-simulator/src/iotdevicesimulator/__assets__/.certs/AmazonRootCA1.pem"
-}
